Android XML Reader from a remote site
This is a android app that demostrates how to read from an xml site feed
mine was 263chat.com . The xml source i would say its not that rich in
Content but for learning reason you can fork or download just to see how
i processed my xwml.

Also feel from to use the code to your benefit.
This was originally made in 2014 . i am not sure if the url for the xml
is still live